Update index.ts
## Additional Functions

### Search by Title or Body

The `searchByTitleOrBody` function allows users to search for todo items based on matching keywords in either the title or body of the todo. This function takes a query string as input and returns a list of todo items that contain the query string either in the title or body. The search is case-insensitive, ensuring that results are not affected by the letter case of the query or todo items. This function provides users with a powerful search capability to quickly find relevant todos based on specific keywords.

### Sort by Date

The `sortByDate` function enables users to sort the todo items based on their creation dates. Users can specify the order of the sorting as either "ascending" or "descending". When the order is set to "ascending", the function returns the todos in the order of oldest to newest creation dates. Conversely, when the order is set to "descending", the function returns the todos in the order of newest to oldest creation dates. Sorting todos by date allows users to easily prioritize their tasks or view the most recent updates. This function enhances the flexibility and usability of the todo management system.